requirements.txt
buttercms-python==1.0
django==2.0.5
